Stats View
Added in stats view basics, will need to be further developed to link in with the database (if do now we cannot run command line at home). Not sure if we only are to show the stats put it above but can add in more/tidy it up.
Also centered the "Top Trump" intro bit